Extract PresentingYourGoalContent component
